+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 All notable changes are documented in this file.
 
-## 0.5.0
+## 0.5.1 - 2020-06-23
+- Fixed all resource links not redirecting to the right url.
+  - Known issue: some resource link cannot be found.
+
+
+## 0.5.0 - 2020-06-22
 Everything is a BandoriLoader now!?
 
 ### Additions